Backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pandas
faiss-cpu
sentence-transformers
langchain
langchain-community
requests
python-dotenv